refactor: modularizar dominios menu e categories
