+++ src/app/routes/dashboard/pages/menu-builder.tsx
@@ -4,7 +4,7 @@ import { useEffect, useMemo, useRef, useState } from "react"
 import { toast } from "sonner"
 import { menuItems } from "@/app/routes/dashboard/data/mock-data"
 import { Modal } from "@/components/ui/modal"
-import { useCategories } from "@/hooks/use-categories"
+import { useCategories } from "@/domains/categories/hooks/use-categories"
 import { useOrganizationCheck } from "@/hooks/use-organization-check"
 import { sentryCaptureException } from "@/lib/sentry"
 import { formatCurrency } from "@/utils/helpers"
