+++ src/app/routes/dashboard/pages/pos.tsx
@@ -16,11 +16,11 @@ import {
 import { useEffect, useMemo, useRef, useState } from "react"
 import { toast } from "sonner"
 import { Modal } from "@/components/ui/modal"
+import { useCategories } from "@/domains/categories/hooks/use-categories"
+import { useProducts } from "@/domains/menu/hooks/use-products"
 import { useCartStore } from "@/domains/pos/store/cart-store"
 import { useRestaurantStore } from "@/domains/restaurant/store/restaurant-store"
-import { useCategories } from "@/hooks/use-categories"
 import { useOrganizationCheck } from "@/hooks/use-organization-check"
-import { useProducts } from "@/hooks/use-products"
 import { sentryCaptureException } from "@/lib/sentry"
 import type { PaymentMethod } from "@/types/dashboard"
 import { formatCurrency } from "@/utils/helpers"
