+++ src/domains/menu/types/menu.types.ts
@@ -1,0 +1,1 @@
+export type { MenuItem } from "@/types/dashboard"
